clean assets prevent img

+++ wordpack/cleanAssets.config.js
@@ -5,10 +5,13 @@ const path = require('path');
 
 fs.readdir('./assets', (err, files) => {
   if (err) console.log(err);
-  console.log(files)
+  // console.log(files)
   for (const file of files) {
+    if (file != "img") {
+
       fs.unlink(path.join('./assets/', file), err => {
-          if (err) console.log(err);
+        if (err) console.log(err);
       });
+    }
   }
 });
